增加URL路径有效性检查并更新端口类型
在httpconfig.go和front.go中增加了对BasePath、ResourcePath、ApiPath等路径有效性的检查。在string.go中添加了IsValidURLPath函数来验证URL路径格式。同时，将mysqlconfig.go和db.go中的端口类型从int更改为int64，并在mysqlconfig.go中增加了端口范围的检查。

+++ src/database/action/internal/db.go
@@ -25,7 +25,7 @@ func ConnectToMySQL() error {
 		config.Config().Yaml.Mysql.DBName)
 }
 
-func connectToMySQL(username string, password string, address string, port int, dbname string) error {
+func connectToMySQL(username string, password string, address string, port int64, dbname string) error {
 	if _db != nil {
 		panic("db is connect")
 	}
